backend/requirements.txt
# 注意：以下版本需要 Python 3.12，请使用 conda 创建 Python 3.12 环境
# conda create -n ivr python=3.12 && conda activate ivr
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic==2.7.4
openpyxl==3.1.5
python-multipart==0.0.9
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
twilio==9.3.2
